first commit
steps 1-3 done.
Write a test to verify the server can issue a random command using a GET request. Then write the functionality that implements this feature.
Connect the client to the server. Using AJAX, request a random swim command from the server. Confirm your swim team is moving around randomly.
Refactor the server program so that instead of responding with random commands, the user's keypress or typed commands are sent in response. Keep your tests passing so be sure to update previously written test(s) to match this change.