commit#4
next level of ajax understanding achieved:

one does not use post requests to make swimmers move.  we use a get from client to server, sending data to the server to do something with it, and get processed data from server back to client to do something else.  in this case, we can issue a command from client side by sending in keypress handler events to the server, the server will process it via httphandler, and send it back for the client to use and do something with it.

for the sake of this project, we refresh the page to process any enqueue'd messages in the messagesQueue, for as long as the data isn't an empty string (processed on client side), and continually fetch for data as long as there is workable data.